Moved the os detection code to a different file so that it can be imported by image resizer
