+++ src/common/OS_Detection.cpp
@@ -1,0 +1,36 @@
+#include "pch.h"
+#include "OS_Detection.h"
+
+// The following three helper functions determine if the user has a build version higher than or equal to 19h1, as that is a requirement for xaml islands
+// Source : Microsoft-ui-xaml github
+// Link: https://github.com/microsoft/microsoft-ui-xaml/blob/c045cde57c5c754683d674634a0baccda34d58c4/dev/dll/SharedHelpers.cpp
+template<uint16_t APIVersion>
+bool IsAPIContractVxAvailable()
+{
+    static bool isAPIContractVxAvailableInitialized = false;
+    static bool isAPIContractVxAvailable = false;
+    if (!isAPIContractVxAvailableInitialized)
+    {
+        isAPIContractVxAvailableInitialized = true;
+        isAPIContractVxAvailable = winrt::Windows::Foundation::Metadata::ApiInformation::IsApiContractPresent(L"Windows.Foundation.UniversalApiContract", APIVersion);
+    }
+
+    return isAPIContractVxAvailable;
+}
+
+bool IsAPIContractV8Available()
+{
+    return IsAPIContractVxAvailable<8>();
+}
+
+bool Is19H1OrHigher()
+{
+    return IsAPIContractV8Available();
+}
+
+// This function returns true if the build is 19h1 or higher, so that we deploy the new settings.
+// It returns false otherwise.
+bool use_new_settings()
+{
+    return Is19H1OrHigher();
+}
