+++ src/common/OS_Detection.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <winrt/Windows.Foundation.Metadata.h>
+
+extern "C" __declspec(dllexport) bool use_new_settings();
