+++ src/runner/pch.h
@@ -29,5 +29,4 @@
 #include <winrt/Windows.ApplicationModel.h>
 #include <winrt/Windows.Storage.h>
 
-#include <winrt/Windows.Foundation.Metadata.h>
 #include <wil/resource.h>
